\name{cab4}
\alias{cab4}
\docType{data}
\title{
Mostly made-up data for testing within-subjects data transformations.
}
\description{
Mostly made-up data for testing within-subjects data transformations.
}
\usage{data(cab4)}
\format{
  A data frame with 32 observations on the following 19 variables.
  \describe{
    \item{\code{Age}}{Subject age in years}
    \item{\code{Ethnicity}}{a factor with levels \code{Caucasian} \code{HIspanic} \code{African American} \code{Other}}
    \item{\code{Gender}}{a factor with levels \code{Male} \code{Female}}
    \item{\code{PC}}{Guilt for property crime, Caucasian suspect}
    \item{\code{PA}}{Guilt for property crime, African-American suspect}
    \item{\code{PH}}{Guilt for property crime, Hispanic suspect}
    \item{\code{DC}}{Guilt for drug crime, Caucasian suspect}
    \item{\code{DA}}{Guilt for drug crime, African-American suspect}
    \item{\code{DH}}{Guilt for drug crime, Hispanic suspect}
    \item{\code{VC}}{Guilt for violent crime, Caucasian suspect}
    \item{\code{VA}}{Guilt for violent crime, African-American suspect}
    \item{\code{VH}}{Guilt for violent crime, Hispanic suspect}
    \item{\code{subj}}{unique identifier}
    \item{\code{anx1}}{Anxiety at time 1}
    \item{\code{anx2}}{Anxiety at time 2}
    \item{\code{anx3}}{Anxiety at time 3}
    \item{\code{dep1}}{Depression at time 1}
    \item{\code{dep2}}{Depression at time 2}
    \item{\code{dep3}}{Depression at time 3}
  }
}
