\name{DeducerReshape-package}
\alias{DeducerReshape-package}
\alias{DeducerReshape}
\docType{package}
\title{
Dialog for Reshaping Data between Wide and Long Formats
}
\description{
Uses a Java GUI dialog to implement the reshape function 
so that users can, for example, convert within-subjects data 
entered in wide format (one subject per row) to long format 
(one observation per row), in preparation for analysis of variance.
}
\details{
\tabular{ll}{
Package: \tab DeducerReshape\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-15\cr
License: \tab GPL-2\cr
}
Overview goes here.
}
\author{
Bill Altermatt <altermattw@hanover.edu>, Helios De Rosario <helios.derosario@gmail.com>
}
